refactor: extract custom footer renderer
